* Query pool refactoring
